chore: remove landing-page code (closes #299)
Deletes the legacy block-based landing-page system and all associated
code. The site-editor (Puck) replaces it as the homepage builder.
