--- src/app/p/[slug]/admin/landing/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/app/admin/properties/[slug]/landing/page';
