+++ src/lib/config/defaults.ts
@@ -20,7 +20,6 @@ export const DEFAULT_CONFIG: SiteConfig = {
   mapStyle: null,
   customNavItems: [],
   setupComplete: false,
-  landingPage: null,
   platformDomain: null,
   puckPages: null,
   puckRoot: null,
